implemented export via clipboard
It's not as clean as a button that sends the json data via email,
but the email plugin was a cordova thing - there's not a way to send
and email from a normal browser.  So this was the best cross-platform
(app and web browser) way to do this as far as I can tell.

Closes #19